Proper Naming Convention of JS & Partial Fix on Model Schema
